Start moving editor code out into src/editor.
