+++ app/src/editor/.eslintrc.js
@@ -1,0 +1,3 @@
+module.exports = exports = {
+    extends: '../userpages/.eslintrc.js',
+}
