+++ app/src/editor/index.pcss
